scripts/one-offs/banner-air-cleanup/requirements.txt
Flask==2.0.1
requests==2.25.1
pytest==6.2.4
dotenv
google.generativeai